testing db for announcement

+++ src/modules/AudreyAnnouncement/AudreyAnnouncement.py
@@ -15,7 +15,6 @@ DB = 'AudreyAnnouncement.db'
 class AudreyAnnouncement(commands.Cog):
 
     def __init__(self, client):
-        self.client = client
         client.permission_authority = self
         self.client = client
         self.cwd = client.config['Bot']['modules_dir'] + 'AudreyAnnouncement/'
@@ -27,19 +26,23 @@ class AudreyAnnouncement(commands.Cog):
 
     @commands.command(help='Clears the announcements of this server')
     async def clearannounce(self, context):
-        conn = sqlite3.connect(self.cwd + DB)
-        server = context.message.guild
-        c = conn.cursor()
-        c.execute('DELETE FROM announcements WHERE server=?', (server,))
-        conn.commit()
-        conn.close()
-        await context.send(f'Announcements cleared for {server.name}.')
+        mod_role = self.client.config['Bot']['mod_role']
+        if mod_role in [role.name.lower() for role in context.message.author.roles]:
+            conn = sqlite3.connect(self.cwd + DB)
+            server = context.message.guild
+            c = conn.cursor()
+            c.execute('DELETE FROM announcements WHERE server=?', (server,))
+            conn.commit()
+            conn.close()
+            await context.send(f'Announcements cleared for {server.name}.')
+        else:
+            await context.send(f'You do not have permission to do that. Ask for the role {mod_role}.')
 
     @commands.command(help="list this server's current announcements")
     async def listannounce(self, context):
         conn = sqlite3.connect(self.cwd + DB)
         c = conn.cursor()
-        results = c.execute('SELECT permission FROM permissions WHERE user=?', (context.message.guild,)).fetchall()
+        results = c.execute('SELECT announcement FROM announcements WHERE server=?', (context.message.guild,)).fetchall()
         announcements = [''.join(row) for row in results]
         announce_string = '\n'.join(announcements)
         await context.send(
@@ -47,7 +50,7 @@ class AudreyAnnouncement(commands.Cog):
         conn.close()
 
     @commands.command(help='Adds announcement to a server')
-    async def addannounce(self, context, announcement, channel: discord.textChannel):
+    async def addannounce(self, context, announcement, channel: discord.channel):
         mod_role = self.client.config['Bot']['mod_role']
         if mod_role in [role.name.lower() for role in context.message.author.roles]:
             conn = sqlite3.connect(self.cwd + DB)
@@ -56,7 +59,7 @@ class AudreyAnnouncement(commands.Cog):
                 channel = discord.utils.find(lambda c: "announce" in c.name or "reminder" in c.name,
                                              context.message.guild.channels)
             c.execute('INSERT INTO announcements(server, channel, announcement) VALUES (?, ?, ?)',
-                      (context.message.guild, channel, announcement))
+                      (context.message.guild.id, channel.id, announcement))
             conn.commit()
             conn.close()
             await context.send(f'Will now announce \'{announcement}\' in {channel}.')
